Action page refactoring, some cleanup
