+++ resources/assets/components/Modal/Modal.js
@@ -27,21 +27,24 @@ class Modal extends React.Component {
 
     // @TODO: These should be injected from outside this component.
     const children = [
-      <AffirmationContainer />,
-      competitionStep ? (
+      <AffirmationContainer key="affirmation" />,
+    ];
+    if (competitionStep) {
+      children.push(
         <CompetitionContainer
+          key="competition"
           content={competitionStep.content}
           photo={competitionStep.photos[0]}
           byline={competitionStep.additionalContent}
-        />
-      ) : null,
-    ];
+        />,
+      );
+    }
 
     return (
       <Portal closeOnEsc isOpened={shouldShowModal}>
         <div className="modal" role="presentation" ref={node => this.node = node} onClick={this.handleOverlayClick}>
           <div className="modal__container">
-            { children.map(child => <div className="modal__slide">{child}</div>) }
+            { children.map(child => <div key={child.key} className="modal__slide">{child}</div>) }
             <button className="modal__exit" onClick={this.props.closeModal}>×</button>
           </div>
         </div>
