--- resources/assets/components/Pitch/PitchCampaigns/data/IncidentsAgainstMuslims.json
@@ -1,35 +1,0 @@
-{
-  "meta": {
-    "title": "Incidents Against Muslims",
-    "axis": {
-      "x": "Year",
-      "y": "Incidents"
-    }
-  },
-  "data": [
-    {
-      "value": 160,
-      "label": "2010"
-    },
-    {
-      "value": 157,
-      "label": "2011"
-    },
-    {
-      "value": 130,
-      "label": "2012"
-    },
-    {
-      "value": 135,
-      "label": "2013"
-    },
-    {
-      "value": 154,
-      "label": "2014"
-    },
-    {
-      "value": 256,
-      "label": "2015"
-    }
-  ]
-}
