--- resources/assets/components/Pitch/PitchContainer/pitchContainer.scss
@@ -1,72 +1,0 @@
-@import '../pitchToolbox';
-
-.pitch-container {
-  width: 100%;
-  display: block;
-  position: relative;
-  min-height: 100vh;
-  overflow-x: hidden;
-
-  &.-half-height {
-    min-height: 50vh;
-  }
-
-  .pitch-container__absolute {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-  }
-
-  .pitch-container__content {
-    position: relative;
-    height: 100%;
-
-    &.-padding {
-      padding: $base-spacing * 2;
-
-      @include media($medium) {
-        padding: $base-spacing * 4;
-      }
-    }
-  }
-
-  .pitch-container__scroll-hint {
-    @include media($medium) {
-      position: absolute;
-      left: 50%;
-      bottom: 15%;
-      width: 64px;
-      height: 64px;
-      margin-left: -32px;
-      display: block;
-      background: url('./scroll.svg') no-repeat;
-      background-size: contain;
-      animation: bounce 4s infinite alternate;
-    }
-  }
-
-  > video {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 200%;
-    height: 100%;
-    object-fit: fill;
-
-    @include media($medium) {
-      width: 100%;
-    }
-  }
-}
-
-@keyframes bounce {
-  from {
-    transform: translateY(0);
-  }
-
-  to {
-    transform: translateY(20px);
-  }
-}
