--- resources/assets/components/Pitch/PitchFlex/pitchFlex.scss
@@ -1,47 +1,0 @@
-@import '../pitchToolbox';
-
-.pitch-flex {
-  height: 100%;
-  display: flex;
-  flex-wrap: nowrap;
-  flex-direction: column;
-
-  @include media($medium) {
-    flex-direction: row;
-  }
-}
-
-.pitch-flex__cell {
-  flex: 1;
-
-  > div {
-    height: 100%;
-  }
-
-  &.-padding {
-    > div {
-      padding: $base-spacing;
-    }
-  }
-
-  @include media($medium) {
-    height: 100%;
-    flex: auto;
-
-    &.-full {
-      flex: 0 0 100%;
-    }
-
-    &.-half {
-      flex: 0 0 50%;
-    }
-
-    &.-one-third {
-      flex: 0 0 33.3333333333333%;
-    }
-
-    &.-two-thirds {
-      flex: 0 0 66.6666666666666%;
-    }
-  }
-}
