--- resources/assets/components/Pitch/PitchHeader/pitchHeader.scss
@@ -1,118 +1,0 @@
-@import '../pitchToolbox';
-
-.pitch-header {
-  height: 100%;
-
-  p {
-    margin-top: $base-spacing * 2;
-    font-family: $primary-font-family;
-    color: inherit;
-    font-size: 18px;
-    width: 50%;
-    margin: $base-spacing auto;
-
-    @include media($medium) {
-      font-size: 24px;
-    }
-  }
-
-  h1 {
-    font-family: $primary-font-family;
-    text-transform: uppercase;
-    color: inherit;
-    background-color: inherit;
-
-    span {
-      padding: 0 $half-spacing;
-      background: $white;
-      color: $black;
-      line-height: 70px;
-    }
-  }
-
-  .button {
-    margin-top: $base-spacing;
-
-    &.-inverted {
-      background: $white;
-      color: #222;
-    }
-  }
-
-  &.-center {
-    > div {
-      text-align: center;
-    }
-  }
-
-  &.-middle {
-    > div {
-      position: relative;
-      top: 50%;
-      transform: translateY(-50%);
-    }
-  }
-
-  &.-bottom {
-    position: relative;
-
-    h1 {
-      position: absolute;
-      bottom: 0;
-    }
-  }
-
-  &.-medium {
-    h1 {
-      font-size: 40px;
-
-      @include media($medium) {
-        font-size: 50px;
-      }
-
-      @include media($large) {
-        font-size: 60px;
-      }
-
-      @include media($larger) {
-        font-size: 70px;
-      }
-    }
-  }
-
-  &.-large {
-    h1 {
-      font-size: 50px;
-
-      @include media($medium) {
-        font-size: 65px;
-      }
-
-      @include media($large) {
-        font-size: 75px;
-      }
-
-      @include media($larger) {
-        font-size: 85px;
-      }
-    }
-  }
-
-  &.-extra-large {
-    h1 {
-      font-size: 60px;
-
-      @include media($medium) {
-        font-size: 75px;
-      }
-
-      @include media($large) {
-        font-size: 85px;
-      }
-
-      @include media($larger) {
-        font-size: 95px;
-      }
-    }
-  }
-}
