--- resources/assets/components/Pitch/pitchToolbox.scss
@@ -1,3 +1,0 @@
-@import '../../scss/next-toolbox';
-
-// This is just for saving me the extra ../../ in all of the files.
